+++ src/Launcher/Config.ts
@@ -26,11 +26,13 @@ export const cerner: LaunchConfig = {
     "pr_questionnaire",
     "pr_questionnaireresponse",
     "pr_servicerequest",
+    "pr_documentreference",
     "pw_allergyintolerance",
     "pw_communication",
     "pw_medicationrequest",
     "pw_observation",
     "pw_questionnaireresponse",
+    "pw_documentreference",
     "ur_allergyintolerance",
     "ur_careplan",
     "ur_communication",
@@ -52,5 +54,6 @@ export const cerner: LaunchConfig = {
     "uw_allergyintolerance",
     "uw_communication",
     "uw_questionnaireresponse",
+    "uw_documentreference",
   ],
 };
